fix(connectdb): Re-use existing DB connection (#21)
This avoids instantiating a new peewee Database (and, hence, creating a new database connection) in the get_peewee_database calls if one already exists. I suspect this is what we intended to happen all along.

Without this fix, every call to chimedb.core.connect() results in a new connection to the database being established. In theory, for normal connections to the MySQL database, this is not a problem: one connection should be the same as any other. (The previous connection is closed during GC).

However, this can cause trouble when using an in-memory SQLite database in test-safe mode.

I also wonder if this is the cause of some of the unexplained DB connection loss we've experienced, especially in alpenhornd, but I have no real evidence for that.