fix(context): Handle SystemExit in atomic decorator (#26)
In normal operation, click command functions don't return, they
call sys.exit(), so we have to catch SystemExit in the decorator
to ensure the transaction is properly exited.

Note: this behaviour only happens when the click CLI is running in
"standalone mode" (which is the normal mode for a click CLI).  When
executed via, say, `CLIRunner` in a unit test, this behaviour is
absent, and the click command will just exit with a return value
like a normal function.